third_party/WebKit/LayoutTests/http/tests/inspector/timeline-xhr-response-type-blob-event-expected.txt

Tests the Timeline events for XMLHttpReqeust with responseType="blob"

XHRReadyStateChange Properties:
{
    children : <object>
    data : {
        readyState : 2
        url : .../inspector/network/resources/resource.php
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    thread : <string>
    type : "XHRReadyStateChange"
}
XHRReadyStateChange Properties:
{
    children : <object>
    data : {
        readyState : 3
        url : .../inspector/network/resources/resource.php
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    thread : <string>
    type : "XHRReadyStateChange"
}
XHRReadyStateChange Properties:
{
    children : <object>
    data : {
        readyState : 4
        url : .../inspector/network/resources/resource.php
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    thread : <string>
    type : "XHRReadyStateChange"
}
XHRLoad Properties:
{
    children : <object>
    data : {
        url : .../inspector/network/resources/resource.php
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    thread : <string>
    type : "XHRLoad"
}
